Find first repeated character
Given a string, find the first repeated character in it. We need to find the character that occurs more than once and whose index of second occurrence is smallest. A variation of this question is discussed here.

An efficient solution is to use Hashing to solve this in O(n) time on average.

Create an empty hash.
Scan each character of input string and insert values to each keys in the hash.
When any character appears more than once, hash key value is increment by 1, and return the character.