First element occurring k times in an array
Given an array of n integers. The task is to find the first element that occurs k number of times. If no element occurs k times the print -1. The distribution of integer elements could be in any range.

Examples:

Input : {1, 7, 4, 3, 4, 8, 7}, 
        k = 2
Output : 7
Both 7 and 4 occur 2 times.
But 7 is the first that occurs 2 times. 

Input : {4, 1, 6, 1, 6, 4}, 
        k = 1
Output : -1
Recommended: Please solve it on “PRACTICE ” first, before moving on to the solution.
Simple Approach: By using two loops. It has a time complexity of O(n2).

Efficient Approach: Use unordered_map for hashing as range is not known. Steps:



 

Traverse the array elements from left to right.
While traversing increment their count in the hash table.
Again traverse the array from left to right and check which element has a count equal to k. Print that element and stop.
If no element has a count equal to k, print -1.